--- pmverifier/snapshot_verifier.go
@@ -1,11 +1,0 @@
-package pmverifier
-
-import (
-	"github.com/vitelabs/go-vite/consensus"
-	"github.com/vitelabs/go-vite/pmchain"
-)
-
-type SnapshotVerifier struct {
-	reader pmchain.Chain
-	cs     consensus.Verifier
-}
