--- verifier/external_reader.go
@@ -1,9 +1,0 @@
-package verifier
-
-import (
-	"github.com/vitelabs/go-vite/ledger"
-)
-
-type Consensus interface {
-	VerifyAccountProducer(block *ledger.AccountBlock) (bool, error)
-}
