--- verifier/metrics.go
@@ -1,3 +1,0 @@
-package verifier
-
-//var forkBranch = metrics.GetOrRegisterHistogram("/verifier_fork", metrics.BranchRegistry, metrics.NewUniformSample(100))
